requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
jinja2
python-multipart
apscheduler
python-jose[cryptography]
passlib[bcrypt]
email-validator
